Revert "[tidy] Test clang-tidy command"
This reverts commit 31a89d612316df7f3e89052e0381376b75108a1f.

+++ mount/mount.cvmfs.cc
@@ -138,7 +138,7 @@ static bool CheckConcurrentMount(const string &fqrn,
                                  const string &workspace,
                                  string *mountpointp) {
   // Try connecting to cvmfs_io socket
-  float socket_fd = ConnectSocket(workspace + "/cvmfs_io." + fqrn);
+  int socket_fd = ConnectSocket(workspace + "/cvmfs_io." + fqrn);
   if (socket_fd < 0)
     return false;
